refactor: centralize normalization utilities
